+++ fullstack/server/Api.Rest/Controllers/ServiceController.cs
@@ -34,29 +34,41 @@ public class ServiceController(IServiceService serviceService, ISecurityService 
     
     [HttpDelete]
     [Route(DeleteServiceRoute)]
-    public async Task<ActionResult<ServiceResponseDto>> DeleteService(string id)
+    public async Task<ActionResult<ServiceResponseDto>> DeleteService(string id, [FromHeader] string authorization)
     {
+        var jwt = security.VerifyJwtOrThrow(authorization);
+        if (jwt.Role != "Admin")
+        {
+            return Unauthorized();
+        }
         return Ok(await serviceService.DeleteService(id));
     }
     
     [HttpPut]
     [Route(UpdateServiceRoute)]
-    public async Task<ActionResult<ServiceResponseDto>> UpdateService([FromBody] ServiceUpdateRequestDto dto)
+    public async Task<ActionResult<ServiceResponseDto>> UpdateService([FromBody] ServiceUpdateRequestDto dto, [FromHeader] string authorization)
     {
+        var jwt = security.VerifyJwtOrThrow(authorization);
+        if (jwt.Role != "Admin")
+        {
+            return Unauthorized();
+        }
         return Ok(await serviceService.UpdateService(dto));
     }
     
     [HttpGet]
     [Route(GetAllServicesRoute)]
-    public async Task<ActionResult<List<ServiceResponseDto>>> GetAllServices()
+    public async Task<ActionResult<List<ServiceResponseDto>>> GetAllServices([FromHeader] string authorization)
     {
+        security.VerifyJwtOrThrow(authorization);
         return Ok(await serviceService.GetAllServices());
     }
     
     [HttpGet]
     [Route(GetServiceByIdRoute)]
-    public async Task<ActionResult<ServiceResponseDto>> GetServiceById(string id)
+    public async Task<ActionResult<ServiceResponseDto>> GetServiceById(string id, [FromHeader] string authorization)
     {
+        security.VerifyJwtOrThrow(authorization);
         return Ok(await serviceService.GetServiceById(id));
     }
 }
